refactor: 검색필터 리팩터링
